\name{datacounts}
\alias{datacounts}
\title{
Internal
}
\description{
Internal. 
}
\usage{
datacounts(response, id, repeated, ncategories)
}
\arguments{
  \item{response}{
response variable.
}
  \item{id}{
id variable
}
  \item{repeated}{
repeated variable.
}
  \item{ncategories}{
categories.
}
}
\value{
A dataframe.
}
\references{
Later on.
}
\author{
Anestis Touloumis
}
\keyword{ internal }

