\name{derivacl}
\alias{derivacl}
\title{
Internal Derivative Matrix For Acl
}
\description{
Internal derivative matrix for acl model.
}
\usage{
derivacl(fitprob, ncategoriesm1, X_mat)
}
\arguments{
  \item{fitprob}{
fitted proportions.
}
  \item{ncategoriesm1}{
ncategories minus one.
}
  \item{X_mat}{
modified design matrix.
}
}
\value{
The derivative matrix of the Acl-model.
}
\references{
Later on.
}
\author{
Anestis Touloumis
}
\keyword{ internal }