\name{derivbcl}
\alias{derivbcl}
\title{
Internal Derivative Matrix For BCL
}
\description{
Internal derivative matrix for BCL model.
}
\usage{
derivbcl(fitprob, ncategoriesm1, X_mat)
}
\arguments{
  \item{fitprob}{
fitted proportions.
}
  \item{ncategoriesm1}{
ncategories minus one.
}
  \item{X_mat}{
modified design matrix.
}
}
\value{
The derivative matrix of the BCL-model.
}
\references{
Later on.
}
\author{
Anestis Touloumis
}
\keyword{ internal }