% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multgee_internals_derivs.R
\name{derivmclm}
\alias{derivmclm}
\title{Derivative Matrix for a Marginal Cumulative Link Model}
\usage{
derivmclm(mueta, ncategoriesm1, X_mat)
}
\arguments{
\item{mueta}{Derivative of linear predictor.}

\item{ncategoriesm1}{number of categories minus one.}

\item{X_mat}{the design matrix}
}
\description{
Derivative Matrix for a Marginal Cumulative Link Model
}
