\name{fitmm}
\alias{fitmm}
\title{
Internal Fitting Function For Marginal Counts
}
\description{
Internal Fitting Function For Marginal Counts.
}
\usage{
fitmm(data, marpars, homogeneous, restricted, add)
}
\arguments{
  \item{data}{
dataframe.
}
  \item{marpars}{
marginal parameters.
}
  \item{homogeneous}{
logical.
}
  \item{restricted}{
logical.
}
  \item{add}{
positive number.
}
}
\references{
Later on.
}
\author{
Anestis Touloumis
}
\keyword{ internal }