\name{inversemat}
\alias{inversemat}
\title{
Internal Inversion Method
}
\description{
Internal Inversion Method.
}
\usage{
inversemat(x, IM)
}
\arguments{
  \item{x}{
a matrix.
}
  \item{IM}{
character.
}
}

\value{
A matrix.
}
\references{
Later on.
}
\author{
Anestis Touloumis
}
\keyword{ internal }