\name{ipfp}
\alias{ipfp}
\title{
IPFP
}
\description{
IPFP function for square tables.
}
\usage{
ipfp(initial, rowmars, colmars, dimension, maxit = ipfp.control()$maxit, tol = ipfp.control()$tol)
}
\arguments{
  \item{initial}{
a vector of the initial values.
}
  \item{rowmars}{
a vector of the row margins.
}
  \item{colmars}{
a vector of the col margins.
}
  \item{dimension}{
an integer indicating the dimensions of the square table.
}
  \item{maxit}{
integer indicating the maximum number of iterations.
}
  \item{tol}{
a positive convergence tolerence value.
}
}
\references{
Later on.
}
\author{
Anestis Touloumis
}
\keyword{ internal }
