\name{LORgee.default}
\alias{LORgee.default}
\title{
The class \code{LORgee}. 
}
\description{
The class \code{LORgee}.
}
\usage{
LORgee.default(x, ...)
}
\arguments{
  \item{x}{
object.
}
  \item{\dots}{
other other arguments.
}
}
\author{
Anestis Touloumis
}
\keyword{internal}