\name{mmpar}
\alias{mmpar}
\title{
Marginalized Margins
}
\description{
Marginalized Margins.
}
\usage{
mmpar(LORem, LORstr, timepairs, homogeneous)
}
\arguments{
  \item{LORem}{
estimation method for LOR.
}
  \item{LORstr}{
working structure for LOR.
}
  \item{timepairs}{
timepairs.
}
  \item{homogeneous}{
logical indicating homogeneous scores.
}
}
\references{
Later on.
}
\author{
Anestis Touloumis
}
\keyword{ internal }
