\name{muprob}
\alias{muprob}
\title{
Transformation To Probabilities
}
\description{
Transformation To Probabilities.
}
\usage{
muprob(cumprob, nobs, ncategoriesm1)
}
\arguments{
  \item{cumprob}{
cumulative probabilities matrix.
}
  \item{nobs}{
number of observations.
}
  \item{ncategoriesm1}{
the category number.
}
}
\references{
Later on.
}
\author{
Anestis Touloumis
}
\keyword{ internal }
