\name{odds.ratio}
\alias{odds.ratio}
\title{
Internal Odds Ratios
}
\description{
Internal Odds Ratios.
}
\usage{
odds.ratio(x)
}
\arguments{
  \item{x}{
A square matrix.
}
}
\value{
An odds ratio matrix.
}
\references{
Later on.
}
\author{
Anestis Touloumis
}
\keyword{ internal }
