\name{print.LORgee}
\alias{print.LORgee}
\title{
Print Functions
}
\description{
Internal.
}
\usage{
\method{print}{LORgee}(x, \dots)
}
\arguments{
  \item{x}{
an object from  class LORgee.
}
  \item{\dots}{
other arguments.
}
}
\author{
Anestis Touloumis
}
\keyword{ internal }

