\name{print.summary.LORgee}
\alias{print.summary.LORgee}
\title{
Summary Print
}
\description{
Internal Function.
}
\usage{
\method{print}{summary.LORgee}(x, \dots)
}
\arguments{
  \item{x}{
an object of the class summery LORgee
}
  \item{\dots}{
other documents.
}
}
\author{
Anestis Touloumis
}
\keyword{ internal }

