% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qic.R
\name{qic}
\alias{qic}
\title{Model Selection Criteria}
\usage{
qic(object, digits = 3)
}
\arguments{
\item{object}{an object of the class "LORgee".}

\item{digits}{numeric indicating the number of decimal points in reported 
summaries.}
}
\value{
Returns a list with the components:
\item{qic}{the QIC criterion.} 
\item{qic_u}{the QIC_u criterion}
\item{cic}{the CIC criterion.}
\item{rjc}{the Rotnumber of observations.}
\item{quisi_likelihood}{the values of the convergence variables.}
}
\description{
add text.
}
\examples{
data(housing)
fitmod <- nomLORgee(y~factor(time)*sec,data=housing,id=id, repeated=time)
qic(fitmod)
}
\references{
Pan, W. (2001) Akaike's information criterion in generalized 
estimating equations. \emph{Biometrics} \bold{57}, 120--125.

Rotnitzky, A. and Jewell, N. P. (1990) Hypothesis testing of regression 
parameters in semiparametric generalized linear models for cluster correlated
data. \emph{Biometrika} \bold{77}, 485--497.
}
\seealso{
\link{ordLORgee} and \link{nomLORgee}.
}
\author{
Anestis Touloumis
}
