\name{RCconstrains}
\alias{RCconstrains}
\title{
Internal Restricted Constraints
}
\description{
Internal Restricted Constraints.
}
\usage{
RCconstrains(ncategories, homogeneous)
}
\arguments{
  \item{ncategories}{
number of categories.
}
  \item{homogeneous}{
logical.
}
}
\value{
A list with various things.
}
\references{
Later on.
}
\author{
Anestis Touloumis
}
\keyword{ internal }
