\name{RRCheter}
\alias{RRCheter}
\title{
Internal Restricted Heterogeneous RC Model
}
\description{
Internal Restricted Heterogeneous RC Model.
}
\usage{
RRCheter(fmla, data, ncategories)
}
\arguments{
  \item{fmla}{
formula.
}
  \item{data}{
data frame.
}
  \item{ncategories}{
category number.
}
}
\value{
Interaction term.
}
\references{
Later on.
}
\author{
Anestis Touloumis
}
\keyword{ internal }
