% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multgee_data.R
\docType{data}
\name{arthritis}
\alias{arthritis}
\title{Rheumatoid Arthritis Clinical Trial}
\format{
A data frame with 906 observations on the following 7 variables:
\describe{ \item{id}{Patient identifier variable.}
\item{y}{Self-assessment score of rheumatoid arthritis measured on a
five-level ordinal response scale.} \item{sex}{Coded as (1) for
female and (2) for male.} \item{age}{Recorded at the baseline.}
\item{trt}{Treatment group variable, coded as (1) for the placebo
group and (2) for the drug group.} \item{baseline}{Self-assessment
score of rheumatoid arthritis at the baseline.}
\item{time}{Follow-up time recorded in months.} }
}
\source{
Lipsitz, S.R. and Kim, K. and Zhao, L. (1994) Analysis of repeated
categorical data using generalized estimating equations. \emph{Statistics in
Medicine}, \bold{13}, 1149--1163.
}
\usage{
arthritis
}
\description{
Rheumatoid self-assessment scores for 302 patients, measured on a five-level
ordinal response scale at three follow-up times.
}
\examples{
data(arthritis)
str(arthritis)
}
\keyword{datasets}
