\name{derivmclm}
\alias{derivmclm}
\title{
Internal Derivative Matrix For MCLM.
}
\description{
Internal derivative matrix for MCLM model.
}
\usage{
derivmclm(mueta, ncategoriesm1, X_mat)
}
\arguments{
  \item{mueta}{
linear predictor.
}
  \item{ncategoriesm1}{
ncategories minus one.
}
  \item{X_mat}{
modified design matrix.
}
}
\value{
The derivative matrix of the Acl-model.
}
\references{
Later on.
}
\author{
Anestis Touloumis
}
\keyword{ internal }


