% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrinsic.pars.R
\name{intrinsic.pars}
\alias{intrinsic.pars}
\title{Intrinsic Parameters Estimation}
\usage{
intrinsic.pars(y = y, data = parent.frame(), id = id, repeated = NULL,
  rscale = "ordinal")
}
\arguments{
\item{y}{a vector that identifies the response vector of the desired
marginal model.}

\item{data}{an optional data frame containing the variables provided in
\code{y}, \code{id} and \code{repeated}.}

\item{id}{a vector that identifies the clusters.}

\item{repeated}{an optional vector that identifies the order of observations
within each cluster.}

\item{rscale}{a character string that indicates the nature of the response
scale. Options include "\code{ordinal}" or "\code{nominal}".}
}
\value{
Returns a numerical vector with the estimated intrinsic parameters.
}
\description{
Utility function to assess the underlying association pattern.
}
\details{
Simulation studies in \cite{Touloumis et al. (2013)} suggested that if the
range of the intrinsic parameter estimates is small then simple local odds
ratios structures should adequately approximate the association pattern.
Otherwise more complicated structures should be employed.

The intrinsic parameters are estimated under the heterogeneous
linear-by-linear association model (\cite{Agresti, 2013}) for ordinal
response categories and under the RC-G(1) model (\cite{Becker and Clogg,
1989}) with homogeneous score parameters for nominal response categories.

A detailed description of the arguments \code{id} and \code{repeated} can be
found in the Details section of \link{nomLORgee} or \link{ordLORgee}.
}
\examples{
data(arthritis)
intrinsic.pars(y, arthritis, id, time, rscale = "ordinal")
## The intrinsic parameters do not vary much. The 'uniform' local odds ratios
## structure might be a good approximation for the association pattern.

set.seed(1)
data(housing)
intrinsic.pars(y, housing, id, time, rscale = "nominal")
## The intrinsic parameters vary. The 'RC' local odds ratios structure
## might be a good approximation for the association pattern.

}
\references{
Agresti, A. (2013) \emph{Categorical Data Analysis}. New York:
John Wiley and Sons, Inc., 3rd Edition.

Becker, M. and Clogg, C. (1989) Analysis of sets of two-way contingency
tables using association models. \emph{Journal of the American Statistical
Association} \bold{84}, 142--151.

Touloumis, A., Agresti, A. and Kateri, M. (2013) GEE for multinomial
responses using a local odds ratios parameterization. \emph{Biometrics}
\bold{69}, 633--640.
}
\seealso{
\link{nomLORgee} and \link{ordLORgee}.
}
\author{
Anestis Touloumis
}
