% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qic.R
\name{qic}
\alias{qic}
\title{Model Selection Criteria}
\usage{
qic(object, digits = 3)
}
\arguments{
\item{object}{an object of the class "LORgee".}

\item{digits}{integer indicating the number of decimal points in reported 
summaries.}
}
\value{
Returns a list with components:
\item{qic}{the QIC criterion.} 
\item{qic_u}{the QIC_u criterion.}
\item{cic}{the CIC criterion.}
\item{rjc}{the Rotnitzky and Jewell criterion.}
\item{quasi_likelihood}{the quasi likelihood.}
}
\description{
add text.
}
\examples{
data(housing)
fitmod <- nomLORgee(y~factor(time)*sec,data=housing,id=id, repeated=time)
qic(fitmod)
}
\references{
Hin, L.Y. and Wang, Y.G. (2009) Working correlation structure identification in generalized estimating equations. \emph{Statistics in Medicine} \bold{28}, 642--658.

Pan, W. (2001) Akaike's information criterion in generalized 
estimating equations. \emph{Biometrics} \bold{57}, 120--125.

Rotnitzky, A. and Jewell, N.P. (1990) Hypothesis testing of regression 
parameters in semiparametric generalized linear models for cluster correlated
data. \emph{Biometrika} \bold{77}, 485--497.
}
\seealso{
\link{nomLORgee} and \link{ordLORgee}.
}
\author{
Anestis Touloumis
}
