\name{summary.LORgee}
\alias{summary.LORgee}
\title{
Summary Model
}
\description{
Internal function.
}
\usage{
\method{summary}{LORgee}(object, \dots)
}
\arguments{
  \item{x}{
a LOR object.
}
  \item{\dots}{
other arguments.
}
}
\author{
Anestis Touloumis
}
\keyword{ internal }

