% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.LORgee.R
\name{vcov.LORgee}
\alias{vcov.LORgee}
\alias{vcov}
\title{Calculate Variance-Covariance Matrix for a Fitted LORgee Object.}
\usage{
\method{vcov}{LORgee}(object, method = "sandwich", ...)
}
\arguments{
\item{object}{a fitted model LORgee object.}

\item{method}{character indicating whether the sandwich covariance matrix
(\code{method = "sandwich"}) or the model--based (naive) covariance matrix
(\code{method = "naive"}) should be returned.}

\item{...}{additional argument(s) for methods.}
}
\value{
A matrix of the estimated covariances between the parameter estimates
in the linear predictor of the GEE model. This should have row and column
names corresponding to the parameter names given by the coef method.
}
\description{
Returns the variance-covariance matrix of the main parameters of a fitted
model LORgee object.
}
\details{
Default is to obtain the robust (sandwich) covariance matrix. However,
\code{robust = FALSE} the naive (model-based) covariance matrix is obtained.
}
\examples{
fitmod <- ordLORgee(formula = y ~ factor(time) + factor(trt) + factor(baseline),
  data = arthritis, id = id, LORstr = "uniform", repeated = time)
vcov(fitmod)

fitmod <- ordLORgee(formula = y ~ factor(time) + factor(trt) + factor(baseline),
  data = arthritis, id = id, repeated = time, LORstr = "uniform")
vcov(fitmod, method = "sandwich")
vcov(fitmod, method = "naive")

}
