% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.LORgee.R
\name{vcov.LORgee}
\alias{vcov.LORgee}
\alias{vcov}
\title{Calculate Variance-Covariance Matrix for a Fitted LORgee Object.}
\usage{
\method{vcov}{LORgee}(object, robust = TRUE, ...)
}
\arguments{
\item{object}{a fitted model LORgee object.}

\item{robust}{whether the robust covariance matrix should be used for
calculating the confidence intervals.}

\item{...}{additional argument(s) for methods.}
}
\value{
A matrix of the estimated covariances between the parameter estimates
in the linear predictor of the GEE model. This should have row and column
names corresponding to the parameter names given by the coef method.
}
\description{
Returns the variance-covariance matrix of the main parameters of a fitted
model LORgee object.
}
\details{
Default is to use robust variance estimates. However, if robust.var is set to
FALSE, naive variance estimates are used.
}
\examples{
fitmod <- ordLORgee(formula = y ~ factor(time) + factor(trt) + 
    factor(baseline), data = arthritis, id = id, LORstr = "uniform",
    repeated = time)
vcov(fitmod)
}
